thinpool: Always update status to match reality
If status ever gets out of sync with reality (for example, if the
cluster reconciler is trying to delete one ThinLV while the node
reconciler is trying to activate another, and both are trying to write
Status at the same time), our code never recovered.  That is because
reconcileThinLvActivations() never updated Status if the actual state
happens to be correct when the Status is stale.  True idempotency
requires that Status should always reflect what we observe in the
system.

Fixes: dd5e5011 ("manager: implement thin LV creation/deletion", v0.5.0)
Signed-off-by: Eric Blake <eblake@redhat.com>