clone: Set label on Volume of node performing data population
The code for populating thin volumes runs in the node reconciler,
which will wake up on every node but should only make progress on one
node, when reconciling the destination Volume.  Right now, the
decision for which node progresses is done by determining whether both
the source and destination LVs have been activated on a given node,
and this can also be observed by the ThinPoolLv object associated with
either source or destination (since a thin volume can only be
activated on one node at a time, there should be at most one node
making progress).  But in the future, we want to also be able to use
dd to populate linear volumes from a thin source, and there is no
ThinPoolLv associated with a linear volume.  It will be easier to
track things if the Volume object itself tracks where the activation
took place; this can be done by temporarily placing a label on the
destination Volume for which node should be doing the work, alongside
the existing label for how to get back to the ThinPoolLv source pool.

Another consideration: in commit e89e54aa ("linear: Use shared
activation for zeroing") it was observed that if the node used for
populating an image (in this case, zeroing it) is not documented in
the CR, but merely depends on what node was running the cluster
controller, then whenever the cluster controller abruptly changes
nodes mid-operation (such as OOMKilled), it can strand an activation
on the old node with no obvious way to track that the old node should
now deactivate.  To fix that, I want to move the block zeroing into
the node reconciler, at which point it is imperative to explicitly
track which node should do the work; but for zeroing, there is no
source ThinPoolLv to refer to.

By doing this, we are now guaranteed that the code that kicks off the
dd process should only be run on one node rather than every node, and
therefore the device paths should exist.

Signed-off-by: Eric Blake <eblake@redhat.com>