linear: Set DataSourceCompleted in node reconciler
Performing blkdiscard as part of CreateBlob in the cluster reconciler
means that the existence of the LV is not advertised in Volume.Status
until the zeroing is complete.  But an upcoming patch wants to change
things to expose LVs to the CSI sidecar as soon as
Conditions["LvCreated"] is true; to do that, the blkdiscard needs to
happen in a separate stage, rather than holding up the CreateBlob
call.  What's more, there is no technical reason why a linear volume
can't be populated by dd from a snapshot source, but for that to
happen, it will be easiest to reuse much of the data population code
for thin volumes which already lives in the node reconciler.

This patch changes the setting of Conditions["DataSourceCompleted"]
for linear volumes out of the cluster controller and into the node
controller, as well as activating at that point (thanks to the label,
we can now track which node has activated, and once again get away
with -aey instead of the -asy needed by commit e89e54aa), although it
will not be until a later patch that the actual zeroing is also moved.
And thanks to the label, it is now easy add in a placeholder to handle
future linear actions and to move the setting of the Conditions to the
common part of the reconcile after either linear or thin has had their
turn at running background job(s).  The shortcut for skipping a trip
through the node reconciler for a thin volume with an empty source is
left in place.

Signed-off-by: Eric Blake <eblake@redhat.com>