crd: Relax constraint on ThinPoolLv.Spec.ActiveOnNode
Prior to the previous patch, the node reconciler really could not
handle the case of Spec.ActiveOnNode set to node B while
Status.ActiveOnNode was still set to node A.  So we fudged around it
by adding a CEL restriction on the CRD that prevents setting
Spec.ActiveOnNode to a different non-empty value if it is currently
non-empty.  But in practice, CEL validations cannot cross subresource
boundaries, and a fast enough client can make two consecutive Spec
edits (first to empty then to a different string) such that the
reconciler which is batching together input events still sees the
problematic scenario.  What's more, because of the CEL restriction,
thinpoollv.go has to jump through hoops to explicitly inactivate
everything and go through some WatchPending to an intermediate state
of Status.ActiveOnNode=="" when an upper layer caller really does want
to switch the active node.

Time to relax the CEL constraint, and in turn add testsuite coverage
that our node reconciler is now robust in the face of a cross-node
swapover with a thinlv that remains in the Active spec category.  This
also adds missing test coverage for activating a thinlv, as well as
for removing a thinlv that is currently activated (most of our code
only attempts removal of an inactive thinlv).  Actually simplifying
thinpoollv.go will be the next commit.

This is a safe ABI change compared to v0.10.0: old CRs still validate
under the new CRD, and mixing the new CRD with old code will still
work.  And by itself, this patch does not add any new code that
violates the old CEL requirements.  But by relaxing this check, it is
imperative that anyone planning to upgrade in-place from v0.10.[01] to
v0.10.2 must first redeploy the CRDs; otherwise, later patches in the
new image may fail to validate under the old CRD.

Signed-off-by: Eric Blake <eblake@redhat.com>