thinpool: Simplify UpdateThinPoolLv
Now that the node reconciler can handle seamless swapover from one
node to another while thinlvs are marked active, there is no longer a
need for the gatekeeper to force an intermediate transition through
inactivating everything.

As mentioned in the previous commit, this patch only works if new
enough CRDs are deployed.

Signed-off-by: Eric Blake <eblake@redhat.com>