release: Capture logs during release script
While preparing for v0.10.2, I hit a failure in Linear mode, but could
not quickly see what the failure was because it had scrolled off my
terminal.

The logs are so long that if something fails, it is worth being able
to grep through the log file rather than having to scroll through a
terminal to decide whether the failure should prevent the release.

Signed-off-by: Eric Blake <eblake@redhat.com>